Make `pubsub-to-big-query-for-sns` a push instead of a pull
Being a `push` subscription means it is called when there is a message
on pubsub topic `sns-events`. As we project that there would be a low
volume of messages on pubsub for the moment, this method was chosen over
a pull because a pull maintains an open connection to pubsub and is
better for streaming data.