Fetch user behaviour after the last flag time for each rule
The previous implementation led to false positives in user behaviour as
it considered the entire behavioural pattern for the user each time
it was being fetched. Thus, whenever a user had been flagged, they would
always be flagged as the user behaviour reflected all user transactions.

This update fixes that, as the user transactions being considered are
those after the `most recent flag time for rule` of a user. Thus only
the transactions after the last flag time are considered fetching the
user behaviour.